PP-13062 handle exception
An unhandled case on our side: we do not anticipate the response to an order inquiry request will give us a
lastEvent for an authorisation error payment but we need to because we could query the order inquiry endpoint if there was a timeout.

Handle <lastEvent>null</lastEvent> Worldpay response case.