Use the new-style switch syntax in GatewayAccountRequestValidator
With the previous commit, we have introduced a new case in the switch
statement validating gateway account requests in order to allow toggling
corporate exemptions.

Since we are touching a switch statement, we are therefore taking this
opportunity to refactor it and to use the new style with arrow syntax rather
than colons and break statements.

Multiple case labels that lead to the same action are now grouped together
using commas.

There's no need for break statements anymore, as each case automatically
breaks after executing its associated action.ß

Further information in Jira.

https://payments-platform.atlassian.net/browse/PP-13038