Authorise with exemption when engine enabled or corporate ex. on
With this change, we are ensuring that the service operates the authorisation
with exemption whether the corporate exemption is set on or also when the
exemption engine is enabled.

Without this change, the service would attempt to authorise with exemption
only when the exemption engine is enabled, missing the setting for corporate
exemptions.

Further information in the PR.

https://github.com/alphagov/pay-connector/pull/5582/files#r1806227518