Clarify test assertions regarding 3ds exemption events
With this change, we are clarifying the tests checking for the
`EXEMPTION_REJECTED` event and the other events.

With a new PR[1], we are now passing instantSource through
WorldpayPaymentProvider so that we can use the timestamp in the equals method
of events in order to assert that an actual event is the same as an expected
event.

In particular we are expecting to see Gateway3dsExemptionResultObtained
events with `EXEMPTION_REJECTED`, `EXEMPTION_HONOURED`, etc.

We are also simplifying the code in various ways for readability.

Further information in the PR[2].

[1]
https://github.com/alphagov/pay-connector/pull/5622

[2]
https://github.com/alphagov/pay-connector/pull/5616#discussion_r1820736340