PP-13052 Test charge update and logging for soft decline
With previous work[1], we have handled the Worldpay authorisation response for
requests with corporate exemptions, with several tests for all the
scenarios.

However, the test for the scenario describing the interaction when the
corporate exemption is soft-declined was not testing that the charge was
being updated and that the logging was taking place.

This change introduces the missing verifications.

Further information in JIRA[2].

[1]
https://github.com/alphagov/pay-connector/pull/5616

[2]
https://payments-platform.atlassian.net/browse/PP-13052?focusedCommentId=58545