PP-13343 fix description for wallet payments
For Worldpay accounts, we have a flag to enable payment references to be sent in the Worldpay order description field instead of the payment description.

However, we are not sending the reference for wallet payments even when the flag is enabled.

For Wallet payments - payment reference should be set in the Worldpay order description if the send_reference_to_gateway flag is enabled.