Introduce new reference attribute in Worldpay Notification
With this change, we are introducing a new reference attribute in the Worldpay
notification, i.e. in our representation of the notification we receive from
Worldpay: refundAuthorisationReference.

This is parsed from the XML element `journalReference` with type
`refund_authorisation`.

We have already been parsing an other reference attribute, parsing it from the
XML element `journalReference` with type `capture`.

An example for the first is <journalReference type="capture"
reference="7500jujg0e8hjap4m9885f8ska"/>. An example for the second is
<journalReference type="refund_authorisation" reference="987654"/>.

This change makes sure that we avoid parsing the wrong reference values, now
that Worldpay will start sending the new reference attribute for successful
refunds.

In the future, it might be good to rename our first internal `reference`
attribute to `captureReference`, to make it clearer that it is parsed from
the journalReference XML element with type `capture`, however we are keeping
it named as `reference`, for now, to limit the scope of this change.

Further information in Jira.

https://payments-platform.atlassian.net/browse/PP-13419