Parse Worldpay notification reference attributes univocally
With this change, we are updating our code responsible for parsing the
Worldpay notifications in order to avoid assigning incorrect values to our
`reference` attribute.

Before recent changes, the parsing would use the first value contained in a
reference attribute of the Worldpay notification assigning it to our
reference property.

With a previous PR[1], we made sure we would be assigning the refund
authorisation reference to a dedicated property first, and then we left the
reference property to be populated with the first value contained in a
reference attribute excluding the refund authorisation reference.

This was not enough, however, because the REFUND_FAILED notification contains
yet another XML element with an attribute named `reference`, which would then
incorrectly be parsed into our `reference` attribute.

With this change we are now univocally assigning the correct reference value,
i.e. using only the value contained in the XML element journalReference with
type `capture` if available.

Additionally, we are parsing the above mentioned reference attribute within
the REFUND_FAILED notification, calling it refundResponseReference.

Note that we have considered using the WrappedStringValue type instead of
String, to enhance type safety, however this cannot be used in this case,
since our values are optional and could be null, which is not an option for
the WrappedStringValue type.

Further information in JIRA[2].

[1]
https://github.com/alphagov/pay-connector/pull/5697

[2]
https://payments-platform.atlassian.net/browse/PP-13419