PP-13419 Handle new successful refund notification
With this change, we are no longer ignoring the SENT_FOR_REFUND notification
sent by Worldpay after a refund submission.

The notification is now included in the REFUND_STATUSES and examined: if it
contains a RefundAuthorisationReference, then it means that the refund
submission was successful, otherwise the notification is ignored.

Since we are now processing different types of notifications arriving after a
successful refund, we are also introducing a check in the Refund Notification
Processor, to avoid setting a refund as successful multiple times, also
avoiding multiple emails to the end user.

Further information in Jira.

https://payments-platform.atlassian.net/browse/PP-13419