PP-13497 Log for Worldpay Refund illegal state transitions
With this change, we are improving the message that will be logged in case of
illegal state transitions for Worldpay refunds to make it clearer.

We are also adding tests to prevent accidental removal or change, given that
this log message is linked to a Splunk Alert.

Further information in Jira.

https://payments-platform.atlassian.net/browse/PP-13497