Temporarily log message to test Splunk Alert
With this change, we are temporarily adding a log message that will be written
only for the Worldpay Test account and only for the SENT_FOR_REFUND
notifications without the refund authorisation reference.

This is needed to test the alert which will be triggered in case of illegal
state transitions.

Further information in Jira.

https://payments-platform.atlassian.net/browse/PP-13497