PP-13445 Parse the reason for a refund failure from the notification
With this change, we are adding the definition for a new attribute in the
Worldpay Notification class in order to use it for the reason for a refund
failure which is included in the journal XML element with journalType
`REFUND_FAILED`.

The attribute is actually the description for notifications for any
journalType.

For this reason, we are introducing the attribute as `description` and then
also adding a dedicated getter than returns the refund failed description for
notifications of type REFUND_FAILED.

This will be logged and then used by the engineer on Support to determine what
to do when a refund request fails.

We will start informing the Support team with a Splunk Alert[1] linked to the
parsed notification.

Depending on the reason for the failure, the Support Team will be able to ask
the Service Admins to either retry the refund (for example for a payment
older than two years) or to find an alternative method outside of Pay
(for example if the credit card had expired).

Further information in Jira[2].

[1]
https://payments-platform.atlassian.net/browse/PP-13497

[2]
https://payments-platform.atlassian.net/browse/PP-13445