PP-13420 Log message for Splunk Alert on REFUND_ERROR
With this change, we are updating the log in the Refund Notification Processor
in Connector, so that a specific message is written when we receive a
notification from Worldpay informing us that a refund request has failed.

This specific message will be used to trigger a Splunk Alert[1], creating a
new support request in Zendesk.

Further information in Jira[2].

[1]
https://github.com/alphagov/cyber-security-splunk-apps/pull/972

[2]
https://payments-platform.atlassian.net/browse/PP-13420