PP-13623 Set transaction ID for Worldpay refunds on creation
For Worldpay, the gateway transaction ID (how Worldpay refers to
the refund) is always the same as the external ID of the refund.
Set the gateway transaction ID as soon as the refund entity is
created, rather than when we process the response to the refund
request to Worldpay, to avoid a possible race condition if we
receive a refund notification from Worldpay immediately.

To do this, introduce a RefundEntityFactory interface so the code
for different payment gateways can supply one with the desired
behaviour.

Co-authored-by: Alex Bishop <alexbishop1@users.noreply.github.com>
Co-authored-by: Hugo Jeffreys <hugo.jeffreys@digital.cabinet-office.gov.uk>
Co-authored-by: Idris Abdirizak <idris.abdirizak@digital.cabinet-office.gov.uk>
Co-authored-by: Marco Tranchino <marco.tranchino@digital.cabinet-office.gov.uk>
Co-authored-by: Sandor Arpa <SandorArpa@users.noreply.github.com>