PP-14114 Fix refund authorisation code casting - version 2
With this change, we are updating the Worldpay Notification Service in order
to accept alphanumeric values sent by Worldpay for the refund authorisation
code when there has been a fast refund authorisation.

Previously we had been accepting numeric values only.

Note that  this change was attempted yesterday,  however there was a bug for
which it treated offline refunds as successful refunds[1], interpreting a
null reference code as an authorisation code.  The bug was quickly reverted
[2].

We are now introducing the functionality again.

[1]
https://github.com/alphagov/pay-connector/pull/5813

[2]
https://github.com/alphagov/pay-connector/pull/5814