PP-14318 Set Worldpay refund as submitted when gateway timeout
With this change, we are addressing a problem that happened recently when the
Worldpay endpoint didn't respond in time to a refund submission request,
leading to Connector setting the refund failed when in fact it had been
correctly submitted.

We have decided to optimistically set the refund as submitted in such cases,
given that we have already a mechanism in place to deal with refunds stuck in
submitted status.

We are also logging a specific message in Splunk in order to be able to count
how many times this happens.