PP-14769 Log calculated states for parity checker mismatches
Some of the more complex parity checker checks involve calculating
two or more named states that represent what the data is in
connector and ledger and then checking if they match according to
specific rules.

When there is a data mismatch caused by the calculated states not
agreeing, add “[calculated_states=…]” to the end of the log line
with a list of the calculated states to make investigation easier.